feat(sessions sample): sanctions flag and presence changes
Added UI toggle for SanctionsEnabled flag for session creation.
Removed presence flag from Session class and session list UI, as presence is a property of the local user and not the session itself.
Removed calls to SetJoinInfo, as it is handled automatically when joining with presence. If not joining with presence, join info probably shouldn't be set either.