feat(android): remove dependency on including gradle files in the published plugin && add support for auto configure eos_dependencies strings.
The string in the eos_values.xml are required for the EOS SDK to
correctly redirect back to the app after directing the user to a browser
to login. Previously, they needed to be modified by the user but actually
couldn't because they were inside the plugin. Now the file is inserted
into a plugin user's Asset's dir, updated on prebuild, and gets merged
when unity does it's apk generation step.