fix(leaderboard sample): Refresh leaderboards after stat ingestion and after wait
EOS-1091; The leaderboards sample had a usability issue where the "Refresh" button unintuitively doesn't refresh the leaderboard rankings. This now refreshes those on press.

After using the UI to "ingest a stat", there is a configurable wait before requerying the leaderboard. This will help with reflecting changes to the leaderboard in a way that feels more responsive. Additionally there is a variable for waiting a number of seconds and requerying after the wait.