fix: EOSManager Instancing
I had originally made the variable that tracks if an EOSManager exists a static boolean. I didn't account for the possibility that all objects are destroyed between automated test runs.

Without this change, after the first test that makes an EOSManager, the subsequent tests will fail to initialize for various reasons.