fix: UIDebugLog enforces the maximum length of a log
If a `Text` element has a `.text` that is sufficiently large, then when trying to render it there will be a failure regarding "too many vertices", capping at 65000. One character does not necessarily map to one vertice. This change imposes a limit of 10000 printed characters. If the log goes over this limit, log messages are dequeued until there is enough room. This number was chosen after some light experimentation; I wasn't able to get a precise value that it exactly needed to be, and this seems to serve well enough.

This does recalculate the "length" of the log cache every time the log is updated. Is that problematic?