feat: Enable PlatformConfigs to Provide Default `IntegratedPlatformManagementFlags`
While migrating, if the old config had no IntegratedPlatformManagementFlags, or the previous value was `Disabled`, try to set the value to a provideable set of default flags. Otherwise, use the old config's IPMFs. This set up currently always provides `Disabled` as the default, but it enables flexibility in platform management of default flags.